requirements.txt
amqp==5.2.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
beautifulsoup4==4.12.3      
billiard==4.2.0
black==23.12.1
bleach==6.1.0
build==1.2.1
CacheControl==0.14.0        
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0       
charset-normalizer==3.3.2   
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
comm==0.2.1
constantly==23.10.4
crashtest==0.4.1
cryptography==42.0.8
daphne==4.1.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.4
django-channels==0.7.0
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-import-export==3.3.5
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fqdn==1.5.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
installer==0.7.0
ipykernel==6.29.2
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keyring==24.3.1
kombu==5.3.7
MarkupPy==1.14
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.4.0
msgpack==1.0.8
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
overrides==7.7.0
packaging==24.0
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.1
platformdirs==4.1.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
Pygments==2.17.2
PyJWT==1.7.1
pyOpenSSL==24.1.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.6
redis==5.0.4
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.13.0
Send2Trash==1.8.2
service-identity==24.1.0
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
stripe==9.3.0
tablib==3.5.0
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4
traitlets==5.14.1
trove-classifiers==2024.7.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.0
waitress==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
whitenoise==6.6.0
widgetsnbextension==4.0.10
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.4.post2